% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPSmultiOmicsSig.R
\name{FPSmultiOmicsSig}
\alias{FPSmultiOmicsSig}
\title{Calculation of ferroptosis-associated molecular signatures}
\usage{
FPSmultiOmicsSig(input_Clinical, input_FPS_score, input_omics)
}
\arguments{
\item{input_Clinical}{data.frame of the clinical information, the column of barcode is samples and the rest column is each clinical feature(For continuous variables, the name could be age_continuous; For discrete variables, the name could be gender_discrete)}

\item{input_FPS_score}{ferroptosis score result from FPS function}

\item{input_omics}{molecular matrix (rownames of the variable must be molecular symbol, each column is a sample)}
}
\value{
data.frame of a propensity score matching (PSM) algorithm result with fdr < 0.05,column of feature.sig is each gene, column of pvalue.sig is p value, column of fdr.sig is fdr, column of coef.sig is coefficient, column of mean0.sig is mean of low group, column of mean1.sig is mean of high group, column of mean0.sig.w is weighted mean of low group, column of mean1.sig.w is weighted mean of high group
}
\description{
A propensity score matching (PSM) algorithm, which appropriately control the effects of clinical confounding factors, was used to calculate ferroptosis-associated molecular signatures between high and low ferroptosis score group.
}
\examples{
data(m3_input_Clinical,package='FPSOmics')
data(m3_input_FPS_score,package='FPSOmics')
data(m3_input_mRNA,package='FPSOmics')
PSM_result=FPSmultiOmicsSig(m3_input_Clinical,m3_input_FPS_score,m3_input_mRNA)
}
