% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPSsurvival.R
\name{FPSsurvival}
\alias{FPSsurvival}
\title{Calculation of ferroptosis score and survival analysis between high and low ferroptosis score groups.}
\usage{
FPSsurvival(input_exp_data, input_survival_data)
}
\arguments{
\item{input_exp_data}{expression matrix (rownames of the variable must be gene symbol, each column is a sample)}

\item{input_survival_data}{data.frame of the survival time information, the name of first colume is barcode which means samples, the name of second colume is OS which means censos status, the name of third colume is OS.time which means survival time}
}
\value{
data.frame of ferroptosis score, column of pro is pro-ferroptosis score, column of anti is anti-ferroptosis score, column of FPS is Ferroptosis score, column of barcode is samples, column of myclusters is high and low group divided by median ferroptosis score; score_MK_plot is MK curve for high and low ferroptosis score
}
\description{
We firstly calculate enrichment score (ES) of ferroptosis-promote genes(pro-FRGs) and ferroptosis-inhibit genes(anti-FRGs) using single sample gene set enrichment analysis (ssGSEA), then ferroptosis score was defined by the ES of pro-FRGs minus anti-FRGs. Secondly, survival analysis between high and low ferroptosis score groups.
}
\examples{
data(m1_input_mRNA,package='FPSOmics') #load mRNA matrix
data(m2_input_survival_data,package='FPSOmics') #load survival time
FPS_survival_result=FPSsurvival(m1_input_mRNA,m2_input_survival_data)

}
